Add segment command which looks in all opened p4 files for a particular regex pattern and moves those to another CL to be handled separately.
This is useful for studios that use change markers to help with integrations, since the change marker can be detected by the regex pattern then those files handled manually, while the rest of the files can be handled through automatic merges.

Also did some python type hints in this same cl, and finished the implemetnation of create_p4_changelist.